[refactor] User/secondTab refactoring 완료
