+++ client/src/Screen/02_SecondTab/Customer/mainScreen.js
@@ -5,8 +5,8 @@ import { createMaterialTopTabNavigator } from '@react-navigation/material-top-ta
 import { objectNullChecking } from '../../../Lib/utils';
 
 // Screens
-import WaitMainScreen from './waitMatching/waitMainScreen';
-import MatchingMainScreen from './processMatching/matchingMainScreen';
+import WaitMainScreen from './wait/waitMainScreen';
+import MatchingMainScreen from './matching/matchingMainScreen';
 
 // Components
 import Loading from '../../../Components/Common/loading';
@@ -14,21 +14,19 @@ import Loading from '../../../Components/Common/loading';
 // Redux Action
 import { getMatchingByCustomerId } from '../../../Contexts/Customer/Matching/action';
 
-function CheckingMatchingScreen({ navigation }) {
+function MainScreen({ navigation }) {
   const { data: user } = useSelector((state) => state.user);
   const { data: matching, loading, error } = useSelector((state) => state.customerMatching);
   const dispatch = useDispatch();
   useEffect(() => {
     dispatch(getMatchingByCustomerId(user.id));
   }, [dispatch]);
-  if (loading || error || !matching) return <Loading />;
-  if (!matching.length) return <WaitMainScreen navigation={navigation} />;
-  else {
-    return (
-      <View style={{ flex: 1 }}>
-        <MatchingMainScreen navigation={navigation} matching={matching[0]} />
-      </View>
-    );
-  }
+  if (loading || error || !matching) {
+    return <Loading />;
+  } else if (!matching.length) {
+    return <WaitMainScreen navigation={navigation} />;
+  } else if (!matching[0].reserved) {
+    return <MatchingMainScreen navigation={navigation} />;
+  } else navigation.replace('Reserved');
 }
-export default CheckingMatchingScreen;
+export default MainScreen;
