+++ client/src/Screen/02_SecondTab/Customer/wait/TabScreen/myProposalScreen.js
@@ -47,7 +47,7 @@ function MyProposalScreen({ navigation }) {
     ]);
   };
   const updateProposal = () => {
-    navigation.navigate('UpdateProposal');
+    navigation.push('UpdateProposal');
   };
   useEffect(() => {
     dispatch(getProposalAsync(user.id));
