+++ client/src/Screen/02_SecondTab/Customer/wait/selectAfterImage/selectFromScrapScreen.js
