+++ client/src/Screen/02_SecondTab/Customer/wait/staticScreen/registeredProposalScreen.js
